DCOS bootstrap script
First pass at the DCOS bootstrap script which downloads all of the major
components in tarball form and extracts them to a versioned DCOS path.
Symlinks are also created so that all resources and executables can be
easily located. For example:

  * /opt/mesosphere/dcos/latest/marathon
  * /opt/mesosphere/dcos/latest/java
  * /opt/mesosphere/dcos/latest/zookeeper

This is currently tested and working with CoreOS build 575 on the Alpha
channel.