Expose JAVA_HOME to the mesos-slave environment
This will be primarily used by the HDFS framework. There is a desire to
have the HDFS framework provide it's own JRE, but that adds a fair
amount of development work since most HDFS end-users will expect HDFS
and the java environment to be globally available on all mesos nodes, so
if we customize it for the HDFS framework, we will need to provide a way
for clients to discover the proper environment to use.