Revert to manually copying shared lib dependencies
Revert to manually copying the shared library dependencies. Attempting
to recursively copy all shared dependencies results in a variety of
other problems, including having to put in place our own ld-linux.so and
override the ELF interpreter for all executables. Leave the copy-libs.rb
script for now in case we decide to revisit this.