Standardize build manifests and improve logging
Standardize all build manifests to be generated in build/*.manifest and
improve logging so that verbose build steps are stored in build/*.log
with annotation of 'stderr: ' or 'stdout: '. We can then send any stderr
results back to stderr so that it is visible to the end-user or build
system.