Remove zookeeper from the mesos pkg dependency
The requires field is intended to reflect both the building and running
of pkgpanda packages. However, in the case of mesos we have a single
package and we don't want to require zookeeper on the mesos slaves. So
we are removing it from the dependency to both simplify the building and
correctly model the running of the mesos package.