Catch all exceptions, ensure fallback written
We must ensure that /etc/resolv.conf *always* has the fallback dns
address written in the event that we are unable to get a list of master
servers from Exhibitor. Otherwise we can end up in a situation where
resolv.conf is pointing to mesos-dns, but mesos-dns is not forwarding to
the fallback DNS server yet because the mesos znode is not yet
initialized. And then mesos core dumps on almost every startup because
it has no DNS server to talk to... :-/

The exception that was missed:
```
An unhandled exception (socket.gaierror: [Errno -3] Temporary failure in
name resolution) caused name resolution to get into a broken state...
```