Dynamically retrieve storage creds for Exhibitor
Switch to using a single Azure Storage Account for both storing VM's as
well as backing Exhibitor. This account is automatically provisioned by
the template.

In order to ensure the storage is provisioned first, we define the
entire cloudconfig.yaml inside the resource group definition, since it
has a dependency on the storage resource being created. That way we can
populate the Azure Storage Key inside the cloud config.

NOTE: The template generation is somewhat broken... to create a new
azuredeploy.json, we have to manually add the listKeys() function after
running `gen_azure_custom_data.py`.