Rework getting active.json during `pkgpanda setup`
Change the behavior so now if active.json is given in setup-flags,
that is used as the active set of packages. As a backup, if
bootstrap-url is set, that will be fetched from next. One of those two
must be specified.

Once active.json has been fetched, get packages from bootstrap-url if
they are not already present locally.