Retry dcos-download until successful
Retry downloading the initial bootstrap tarball until it has succeeded
at least once. This fixes: DCOS-1416 and makes slaves continue to retry
the download, even if the NAT gateway is not yet available (which
happens when provisioning slaves and the NAT instance in parallel with
CloudFormation).

Switch to wget for the changed error handling behavior.  With curl a 403
is always treated as a success, but with wget it will exit with an error
code. We always want to retry on 403's because Amazon may return this if
the resource is temporarily unavailable.