Add SASL authentication to mesos
The plugin directory is now being copied over in addition to the main
shared library.

Since the plugin directory is in a non-standard location, the SASL_PATH
variable needs to be set to indicate the path to the plugins on both
the master and slave hosts.

Mesos is now compiled with the --with-sasl flag so it can make use of
the SASL library