Support custom attributes and resources
This is a patch for Verizon which supports customizing the mesos-slave
attributes and resources via an EnvironmentFile:

    /var/lib/dcos/mesos-slave-common

Contents should be set before the first launch of the slave agent,
otherwise mesos-slave will fail to start due to a conflict with the
initialized slave agent state.

This is a partial implementation of the final solution which will go
into the master branch and is tracked by DCOS-2606.