Allow traffic from Azure management IP
The Microsoft Azure platform uses a static, publicly routable IPv4
address for a variety of administrative scenarios. The IP address is
168.63.129.16. This IP address should not be blocked by any firewalls,
because it can cause unexpected behavior. With respect to Azure Internal
Load Balancing, this IP address is used by monitoring probes from the
load balancer to determine the health state for virtual machines in a
load balanced set.