Update defaults and README
Add the mesosphere-demo SSH key as a default, which we can share with
Microsoft. The private key is available on Onelogin. Set the number of
masters to 1, since multi-master HA is not working as of yet and to keep
cluster sizes reasonably small for now. Limit the number of private
slaves to 1-40. Setting the number of private slaves to 0 will cause a
cluster deployment failure and creating with more than 40 nodes will
almost certainly hit quota and Azure storage limitations.