Allow specifying the name of the build script explicity
This becomes critical when having multiple build variants of a package
and wanting to share build scripts (or explicitly not).

Only variants are allowed to change the build_script though because a
folder is only considered a package if it contains a build script.
As a result, if someone has a default build without a build script it
is likely to cause bizzareness in a `mkpanda tree`. To prevent this,
the default package varant which is always built is required to not
use a custom named build_script