Allow depending on a package variant.
Dependencies are given in the "requires" array in buildinfo.json. They
can be specified by giving just a string, meaning the "default" build
of that package, or by giving a dictionary which contains both the
package name and variant.

Ex:

{
	"requires": [
			"foo",
			{
				"name": "bar",
				"variant": "baz"
			}
		]
}

TODO / WIP: Depending on a different set of packages in a varint from
the base package may cause things to not get built in the right order.