Allow having an "extra" folder inside of a pacakge that will be included in a built package
@lingmann suggested this a while ago. We're getting more and more
packages of little tiny scripts, this makes it so we shouldn't need
another repository for those packages.

Everything inside the "extra" folder is included in the buildids,
regardless of whether or not it is in the gitignore.

Finding the id of everything in the extra folder is surprisingly
difficult. To do it with git, everything would need to be committed
which is a bit painful from a workflow POV. We also wouldn't want
gitignored files to break the "repeatable build" notion. Ideally there
would be no untracked files in the folder though, as tools dropping
them there would make it hard to reproduce builds.