Simplify input, remove region parameter
Make it so there is only one way to set the resourcegroup region. When
submitting an ARM template, it is required to specify the region. By
having the region as a parameter, we run the risk that someone will
specify a different value for the region vs what was submitted to the
Azure API at template creation time.

This also fixes the problem where a user will have to set the region
twice when launching an ARM template from the click-to-deploy button
(once as a parameter and once as one of the required input options for
all templates.) If the value was not set consistently, the template
would fail to launch.

Now we always detect the location using the `resourceGroup()` reference.