Disable hostname lookup on Mesos slaves
Setting hostname_lookup to false will cause Mesos slaves to always
register with the hostname set to the IP address (which is determined by
MESOS_IP_DISCOVERY_COMMAND=/opt/mesosphere/bin/detect_ip). In
environments where hostname lookup and DNS is unavailable (Azure,
on-prem environments, etc.) this causes problems with a number of our
frameworks, since they often depend on the mesos hostname in the
resource offer to be reachable. This change ensures that the hostname in
the framework offer is always set to the IP address as reported by the
ip detect script.