Fix DCOS management IP on Azure
Workaround for the problems around the cluster management IP that is
exposed in the DCOS UI and used by the DCOS CLI.

On Azure COREOS_PUBLIC_IPV4 is exposed as an env var only when nodes are
connected to a load balancer, and in this case, the IP exposed is the
load balancer IP. This value is used for SSH access, when can be
problematic when multiple master nodes are behind the load balancer.

Fix the issue for the time being by:

1. Restrict Azure clusters to a single master node instance. This is
   necessary for multiple reasons, see: DCOS-3329

2. Forward SSH traffic over the load balancer endpoint

This results in a SSH'able COREOS_PUBLIC_IPV4 exposed on the master
node, which is then picked up by adminrouter.