Update to support tox 2.0+
Tox versions 2.0+ have some behavior changes in terms of ssh-agent. In
order to pass along the ssh-agent environment with newer versions of
Tox, a passenv option must be specified in tox.ini. See:
https://bitbucket.org/hpk42/tox/issues/253/newer-tox-prevents-usage-of-ssh-agent-201

This should fix the issues we are seeing with our coding standards test
fail on newer Teamcity agents (agent01 and agent02 were based on Ubuntu
14.04 and Tox v1.9.2). Ubuntu 15.04 ships with tox v2.1.1.