Workaround to support more frameworks on Azure
Various DCOS universe packages depend on local hostname resolution working
(e.g. `ping $(hostname)` must work). This fix makes it so that the local agent
hostname is resolvable to the private ip of the instance.

This config-writer service can be removed when we are certain that all DCOS
packages work properly in environments where hostname DNS resolution is not
functional. See: DCOS-3719