Remove unused save_user_config, save_final config
Motivated by the extracting of config reading / writing to the genconf
library. The save_user_config should have a round trip guarantee
(What gets written can be read back in) but didn't actually test it.
With moving to YAML as the config format, would have to teach genconf
about YAML, or make it the callers responsibility. Doing the latter,
since the caller is learning how to load the config as well.