Teach providers to extract the package files they need into a cloud-config
Previously they wrote them out as package files inside of the cloud-config.
Now they include them in the dcos-config package which is going to be the
only DCOS configuration package, then pull out the files they need into
the cloud-config by explicitly listing them.

This allows the configuration to be shared between all the providers since
the config looks the same for both onprem (not pulled out) and cloud (pulled out)

Writes a pkginfo.json into the cloud-config setup-packages automatically
so that it is always there. Stealing a pkginfo.json from dcos-config
package would result in that package not having one and thus not being
loadable.