Add curl package to DCOS
Build DCOS components with our own bundled libcurl. This is necessary
for Ubuntu systems which does not include libcurl by default, and
reduces our dependency on system libraries.

Build flags are set based on features currently enabled/disabled in the
CoreOS build along with some options to help with debugging and
performance.