Download before doing the local pip wheel building
This makes it so that we save quite a bit of time re-downloading the
same things / checking an external source that things are identical.

Also add wheelhouse to gitignore so it doesn't show up for people.

Note it is a hard failure that wheelhouse not exist before prep_teamcity
is run, because the artifact cleaner needs to clean it up between runs
otherwise we can't guarantee we're getting the right version of the
dcos-image and pkgpanda wheels out of it.