The dcos-signal shouldn't be always auto re-started, just triggered by timer
This was causing us to DDOS segment by pinging once every 5 seconds from every
machine in every DCOS cluster

In addition, since it runs only once an hour, it needs to succeed even when the
mesos masters aren't up otherwise the cluster health check will fail (And it
won't become healthy until after an hour). To get this, teh pinging of
leader.mesos is moved into the service, the service logs whether it reported
analytics or not. Failing to find leader.mesos isn't considered a failure of the
analytics service