Make --deploy and friends unusable unless bootstrap_url is set to file:///opt/dcos_install_tmp
In the gen library shipping with DCOS 1.5 there is no way to say "if this parameter is set then, and
only then, ensure this is set this way".

That means we can't look at the config given to --genconf and only if "ssh_deploy" is selected
default bootstrap_url to file:///opt/dcos_install_tmp. Instead, what we do is ensure if you call the
deploy library it is set properly.