Remove implies, enable using switches inside of templates
This does a minimal removal of implies, making it so that we can evaluate parameters more directly.

This also lines up "conditionals" and mixins to be very, very similar structurally (read: They'll be
merging soon). We're also getting more of a "qualified" graph of different parameters / arguments
(x only exists if y), although it is intentionally depth limited in some places for ease of
implementation.