Move to a unified "setter" representation for nodes in calc graph
This moves all the code that was dealing with "arguments" vs. "defaults" vs. "must" vs. "can" vs. manual set additional things. It also makes "conditionals" handled at the same time as all the other top level configs.

They are then all resolevd at the same time, in one big lazy evaluation of the graph to get the set of parameters we need to actually fill in the holes in the templates.

This also removes --interactive mode for now.