multiple chain support for run_command_chain_async
we need a way to execute several chains at once per host.
A good use case is cleaning up after main chain executed. Because of async
nature, and the way we use if from aiohttp.web, we cannot block the POST
request, wait for it to finish and then execute the cleanup. This PR will add
support of multiple chains execution per host. Now we can pass a list of
CommandChain() instances and they will be executed sequentially without
blocking POST request.

Also this PR will remove `host_status_count` field. This hardcoded value is not
used by FE. And it would make sense to calculate it on the fly based on `hosts`
values.