Switch docker to use overlayfs as it's storage backend.
By default it was using devicemapper. The devicemapper would try to find free
space on the drive, and then make a volume. That would fail, and it would fall
back to devicemapper with a loopback device.

Devicemapper with a loopback device is very heavily argued against by everyone
in the docker community.

In addition, when we did a systemctl restart systemd-journald, that would cause
lvm to restart which would break the whole setup, causing the integration test
to fail.