Remove retry, switch py.test to giving structured results, and set logLevel to debug in ccm test
integration_test.py already retries internally until the hosts are up. If the tests fail they aren't
going to pass magically on the second run. The re-running also messes up TeamCity picking up test
results.

setting logLevel to debug makes it so we see what commands are being run when, which can be useful for debugging.