Support SSHing to master nodes via Master LB
Taking some ideas from the ACS template, we can support SSH'ing to the
master nodes via a static NAT mapping of ports starting at 2200 for each
master node. For example, the first master will always be reachable at
2200, the second at 2201, and so on.

This makes it easy to deterministically SSH to master nodes in the
cluster via a single LB endpoint, and it solves the issue of SSH host
keys changing out from under the user.