Make deploy wait on leader.mesos for master nodes
Attach a customScriptExtension to the master nodes which simply waits
for leader.mesos to respond. This is similar to the dcos-signal service,
but adapted for Azure.

Ensures that the ACS template deploy is not reported as COMPLETE to the
end-user until the master nodes have established quorum and leader.mesos
is resolvable. This also means the DCOS web ui will be available for the
end-user at this point.

It might be desirable to implement this as a dcos-signal service... but
only if we want to consider dcos-signal as both a push and a pull
service.  Currently with AWS dcos-signal is checking leader.mesos and
then pushing an alert to AWS when the cluster has converged. For Azure,
we most poll/block via customScriptExtension, and it seems overkill to
poll a dcos-signal service when we can accomplish the same goal using
just customScriptExtension instead of customScriptExtension + some new
dcos-signal polling service.