(qe-392) improve logging around invalid json
We're experiencing an issue with the automated dcos-image builds where
CCM isn't always returning invalid JSON. But unfortunately, I'm unable
to reproduce the issue locally.

This commit adds debug level logging and outputs the raw response from
the CCM API should an error occur so that it will appear in the build
logs in TeamCity. This way we can gather more data.