Update Azure templates to Ubuntu 16.04 Daily Build
* Switch to Ubuntu 16.04 Daily Build
* Update to Docker 1.11.0-rc2 testing build for Xenial
* Remove apt-get update, it is not necessary and adds an external
  dependency

This update switches to Ubuntu 16.04 and also includes a fix for the
DCOS timer units (setting OnBootSec timeout to a non-zero value) to
prevent Systemd from crashing. The underlying cause for this has been
fixed upstream with the Canonical package, but it has not yet landed in
the daily Xenial builds. Once this fix lands in the Daily builds the
OnBootSec value could be changed back to 0. See:
https://bugs.launchpad.net/ubuntu/+source/systemd/+bug/1554861

FIXES: DCOS-4267
FIXES: DCOS-5410