Provision ACS with fixed number of public agents
The Azure managed ACS service has only a single variable exposed for
controlling the number of agents and this variable has been used up
until now to control both the number of public and private agents (not
ideal).

Instead, we will use this variable to just manage the number of private
agents. The number of public agents will be pre-determined based on the
number of master nodes selected.

For storage, we create a single public agent storage account in order to
isolate it from the other cluster elements. A single storage account
should be sufficient given that we will only see 1 or 3 public agents.

FIXES: DCOS-5400