Add a mesos agent on the masters, so tasks can be launched on all hosts via marathons
Has a new role to prevent tasks from being added their accidentally.

The master intentionally has a small available port range, we run a ton of things on the master and are likely to add more. We don't enable the network isolator so people can bind to ports mesos didn't offer if they wish, should bypass Mesos' offer ports.

In time the goal is that more and more services currently on the masters get taught to be run by Mesos rather than directly on the masters, and then the port range should be expanded