Azure integration test timeout adjustments
Adjust timeouts to compensate for slow Azure deployments. Over the last
38 builds we've had 19 fail to deploy in the 15m window given. The
median time for a successful deploy during this window of failures was
only 4min, but there was 1 successful deploy which took the full 15min.
Of the 38 builds, 7 failed to delete in time, so be much more forgiving
on deletes as well.

The adjustments are:
* wait for deploy: 15m => 30m
* wait for delete: 20m => 60m