Allow usage of binary artifacts again
This allows using pre-built wheels which should speed up docutils install considerably.

These were disabled because when the wheelhouse was shared with the installer docker, the Python host string would change between the host and inside the docker, so only non-built artifacts could be used. Now these are only used on the build and test hosts, which they should always be the right version for