pkgpanda: Add a "source fetcher" abstraction
Consolidates all the code for dealing with a specific kind of source (such as git) into one location.

This is also the first step in a multi-part refactor which should fix the "git_local` "reference is not a tree" bug, as well as speed up cached builds on machines which don't have a full "cache" directory (clean sources was used).