Reduce probability of ACS deploy failure
Reduce the probability of ACS templates from failing to deploy due to
issues with the customScriptExtension (see: DCOS-5728). Currently, a
customScriptExtension can fail if there is moderate load on the host,
irrespective of what the customScriptExtension is doing. By restricting
this to 1 instance and the last master node in the cluster we will
reduce the likelihood that we will hit the issue.

Restore ping output since it is helpful from a logging perspective.