Serialize yaml `true` / `false` back to strings as lower case 'true' / 'false'
[YAML is a typed language](http://www.yaml.org/spec/1.2/spec.html). Things like
boolean, list, null, dictionary are all defined and according to the spec should
map to the language builtin type of the same name (10.2.1.2). In python, this
means `true` and `false` turn into the python Boolean type. Python, when
printing its boolean type by default prints 'True' or 'False' because when
writing python `True` and `False` are the identifiers for true/false. All the
code in gen however checks for lower case 'true' / 'false'.

Update the installer when reading a yaml configuration file to convert the
python boolean into a lower case string